==> covid_situacaoBrasil/P01-covid_mineracao_dados.ipynb <==
# Projeto Covid - Situação Brasil
## Por: Michel Souza Santana
> Data: 20/01/2023

### Objetivos
* Requisitar de um site on-line via url dados sobre a covid19
* Armazenar os dados capturados em uma tabela
* Organizar e limpar os dados

> O propósito final desse projeto é fazer uma comparação entre casos confirmados e mortes registradas para avaliar a letalidade desta doênça.

OBS: A parte de análise desse projeto segue no arquivo anexo 'PS2-covid_analise_dados.ipynb'

> Importação das bibliotecas a serem utilizadas no projeto

import pandas as pd
import numpy as np
from urllib.request import urlopen
from bs4 import BeautifulSoup

> Importando dados via URL do site https://news.google.com/covid19/map?hl=pt-BR&mid=%2Fm%2F015fr&gl=BR&ceid=BR%3Apt-419 (Escolhido aleatoriamente)

* Após importação dos dados, estarei fazendo uso do BeaultifulSoup para enchugar estes por meio de seleção via 'tag'

html = urlopen('https://news.google.com/covid19/map?hl=pt-BR&mid=%2Fm%2F015fr&gl=BR&ceid=BR%3Apt-419')
bs = BeautifulSoup(html.read(), 'html.parser')
dados = bs.findAll('tr', {'class': 'sgXwHf wdLSAe YvL7re'})
dados[:2]

> Nesse passo estarei separando os dados e criando a tabela (DataFrame)

new_data = []
for d in dados:
    data = []
    td = []
    data.append(d.find('div', {'class': 'pcAJd'}).get_text())
    td = d.findAll('td', {'class': 'l3HOY'})
    data.append(td[0].get_text())
    data.append(td[1].get_text())
    data.append(td[3].get_text())
    data.append(td[4].get_text())
    new_data.append(data)
new_data[:5]

> Criando o Data Frame

df = pd.DataFrame(new_data, columns=['local', 'total_casos', 'novos_por_dia', 'a_cada_milhao', 'total_mortes' ])
df

> Percebe-se que na coluna 'a_cada_milhao' existem dados escritos e não numéricos. Tratando e substituindo por '0'

df['a_cada_milhao'] = df['a_cada_milhao'].str.replace('Não há dados', '0')
df

Na coluna novos por dia não há informações relevantes, estarei eliminando esta coluna

df = df.drop(columns=['novos_por_dia'])
df

df.dtypes

> Presiso converter os dados numéricos para int64 e para não haver erros se faz necessário eliminar os caracteres '.'

df['total_casos'] = df['total_casos'].str.replace('.', '')
df['a_cada_milhao'] = df['a_cada_milhao'].str.replace('.', '')
df['total_mortes'] = df['total_mortes'].str.replace('.', '')

> Caracteres '.' eliminados

df

Convertendo os dados para 'int64'

df['total_casos'] = df['total_casos'].astype('int64')
df['a_cada_milhao'] = df['a_cada_milhao'].astype('int64')
df['total_mortes'] = df['total_mortes'].astype('int64')
df.dtypes

> Limpeza e preparação do DataFrame finalizadas salvando em formato csv

df.to_csv('csv/casosXmortes-covid.csv')

df
