==> covid_situacaoBrasil/P02-covid_analise_dados.ipynb <==
# Projeto Covid - Situação Brasil
## Por: Michel Souza Santana
> Data: 20/01/2023

> O propósito final desse projeto é fazer uma comparação entre casos confirmados e mortes registradas para avaliar a letalidade desta doênça.
> Após alcanças o resultado com os estados com maior número de caso e mortes (letalidade), o próximo passo será analisar o índice de vacinação alcançada por estado e em seguida determinar se as vacinas de fato auxiliaram no tratamento ou não. (Projeto P03-covid_vacinas.ipynb)

### Análise dos dados
* Letalidade
* Taxa de mortalidade

> Importação das bibliotecas a serem utilizadas no projeto

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

df = pd.read_csv('csv/casosXmortes-covid.csv')
df.drop(columns=['Unnamed: 0'])
df_populacao = pd.read_csv('csv/populacao.csv')

> Verificando Quntidade de colunas e linhas

df.shape

> Verificando os tipos de dados existentes na tabela

df.dtypes

> Calculando a Letalidade da Doença por região

* A letalidade mede a severidade de uma doença e é definida como a proporção de mortes dentre aqueles doentes por uma causa específica em um certo período de tempo.

df.drop(columns=['Unnamed: 0'], inplace=True)
df

df.shape

df = df.sort_values(by=['local'])
df = df.reset_index()
df.drop(columns=['index'], inplace=True)
df

df = df.join(df_populacao)
df.drop(columns=['Unnamed: 0', 'estado'], inplace=True)

df

df['letalidade %'] = df['total_mortes'] / df['total_casos'] * 100
df

df['mortalidade %'] = (df['total_mortes']/df['populacao']) * 100
df

for i in range(len(df['a_cada_milhao'])):
    if df['a_cada_milhao'][i] == 0:
        df['a_cada_milhao'][i] = df['a_cada_milhao'].mean()

df

df['letalidade %'] = df['letalidade %'].round(2)
df['mortalidade %'] = df['mortalidade %'].round(2)

> Salvando uma cópia da tabela comtodos os dados para uso futuro

df.to_csv('csv/estados.csv')

corr = df.corr()
corr.style.background_gradient(cmap='coolwarm').set_precision(2)

df = df.sort_values(by=['letalidade %'], ascending=False)
letalidade = df.head(10)
letalidade

df = df.sort_values(by=['mortalidade %'], ascending=False)
mortalidade = df.head(10)
mortalidade

df = df.sort_values(by=['total_casos'], ascending=False)
casos = df.head(10)
casos = casos.reset_index()
casos.drop(columns=['index'], inplace=True)
casos 

df = df.sort_values(by=['total_mortes'], ascending=False)
mortes = df.head(10)
mortes = mortes.reset_index()
mortes.drop(columns=['index'], inplace=True)
mortes 

df = casos.merge(mortes, left_on='local', right_on='local', suffixes=('', '_'))
df

df = df.drop(columns=['total_casos_', 'a_cada_milhao_', 'total_mortes_', 'populacao_', 'letalidade %_', 'mortalidade %_'])
df

Informações do resultado validadas no site: https://www.cnnbrasil.com.br/nacional/18-estados-brasileiros-registram-mais-mortes-por-covid-19-em-2021-do-que-em-2020/

df.to_csv('csv/estados_mais_afetados.csv')
