==> covid_situacaoBrasil/P03-covid_vacinacao.ipynb <==
# Projeto Covid - Situação Brasil
## Por: Michel Souza Santana
> Data: 21/01/2023

> Após alcançar o resultado com os estados com maior número de casos e mortes (letalidade), o próximo passo será analisar o índice de vacinação alcançada por estado e em seguida determinar se as vacinas de fato auxiliaram no tratamento ou não. (Projeto P03-covid_vacinas.ipynb)

### Análise dos dados
* Índice de vacinação

> Iniciando as bibliotecas python a serem utilizadas no projeto

import pandas as pd
import numpy as np
from bs4 import BeautifulSoup
from urllib.request import urlopen
import matplotlib.pyplot as plt

> Captura dos dados referente às vacinações realizadas em cada estado: https://especiais.g1.globo.com/bemestar/vacina/2021/mapa-brasil-vacina-covid/ 

> As vezes capturar dados direto do site se torna um pouco complexo e quando se trata de um baixo volume, podemos captura-lo direto do código fonte da página, exporta=lo para um arquivo e então carrega-lo no projeto.

dados = pd.read_html('html_vacinacao.html')
df = pd.DataFrame(dados[0])
df

df['2ª doseou doseúnica(Unidades)'] = df['2ª doseou doseúnica(Unidades)'].str.replace('.', '')
df['2ª doseou doseúnica(Unidades)'] = df['2ª doseou doseúnica(Unidades)'].astype('int64')

df['2ª doseou doseúnica(Em %)'] = df['2ª doseou doseúnica(Em %)'].str.replace(',', '.')
df['2ª doseou doseúnica(Em %)'] = df['2ª doseou doseúnica(Em %)'].str.replace('%', '')
df['2ª doseou doseúnica(Em %)'] = df['2ª doseou doseúnica(Em %)'].astype(float)

df['Dose dereforço*(Unidades)'] = df['Dose dereforço*(Unidades)'].str.replace('.', '')
df['Dose dereforço*(Unidades)'] = df['Dose dereforço*(Unidades)'].astype('int64')

df['Dose dereforço*(Em %)'] = df['Dose dereforço*(Em %)'].str.replace(',', '.')
df['Dose dereforço*(Em %)'] = df['Dose dereforço*(Em %)'].str.replace('%', '')
df['Dose dereforço*(Em %)'] = df['Dose dereforço*(Em %)'].astype(float)

df['1ª dose(Unidade)'] = df['1ª dose(Unidade)'].str.replace('.', '')
df['1ª dose(Unidade)'] = df['1ª dose(Unidade)'].astype('int64')

df['1ª dose(Em %)'] = df['1ª dose(Em %)'].str.replace(',', '.')
df['1ª dose(Em %)'] = df['1ª dose(Em %)'].str.replace('%', '')
df['1ª dose(Em %)'] = df['1ª dose(Em %)'].astype(float)

df['Total devacinasaplicadas'] = df['Total devacinasaplicadas'].str.replace('.', '')
df['Total devacinasaplicadas'] = df['Total devacinasaplicadas'].astype('int64')

df = df.sort_values(by=['Total devacinasaplicadas'], ascending=False)
df

df = df.reset_index()
df.drop(columns=['index'], inplace=True)
df

df = df.rename(columns={'Estado': 'estado', 
                   '2ª doseou doseúnica(Unidades)': '2ª dose', 
                   '2ª doseou doseúnica(Unidades)': '2ª dose',
                   '2ª doseou doseúnica(Em %)': '2ª dose %',
                   'Dose dereforço*(Unidades)': 'reforco',
                   'Dose dereforço*(Em %)': 'reforco %',
                   '1ª dose(Unidade)': '1ª dose',
                   '1ª dose(Em %)': '1ª dose %',
                   'Total devacinasaplicadas': 'total vascinas'
                   }) 
                     

df.to_csv('csv/vacinacao.csv')
